hasdvjsa<dfaf

+++ Flow_1/REST_Persons/src/main/java/facades/IPersonFacade.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface IPersonFacade {
     public PersonDTO add(PersonDTO personDTO);
-    public PersonDTO delete(int id) throws PersonNotFoundException;
-    public PersonDTO get(int id);
+    public PersonDTO delete(Long id) throws PersonNotFoundException;
+    public PersonDTO get(Long id);
     public List<PersonDTO> getAll();
     public PersonDTO edit(PersonDTO personDTO) throws PersonNotFoundException;
 }
